requirement.txt
streamlit
seaborn
matplotlib
urlextract
pandas
numpy
wordcloud
emojis
stopward.txt
